fix: Unity 6000.5 EntityId migration
- Replace Object.GetInstanceID() with (long)EntityId.ToULong(obj.GetEntityId())
- Replace EditorUtility.InstanceIDToObject(int) with EditorUtility.EntityIdToObject(EntityId.FromULong((ulong)id))
- Change all instanceId parameters from int? to long?
- Update JSON parsing from ToObject<int?> to ToObject<long?>

Unity 6000.5+ renamed InstanceID APIs to EntityId. Instance IDs are now 64-bit.

+++ Editor/Models/UpdateGameObjectRequest.cs
@@ -7,7 +7,7 @@ namespace McpUnity.Models
     public class UpdateGameObjectRequest
     {
         [JsonProperty("instanceId")]
-        public int? InstanceId { get; set; }
+        public long? InstanceId { get; set; }
 
         [JsonProperty("objectPath")]
         public string ObjectPath { get; set; }
